backend/requirements.txt
Flask==1.1.2
redis==3.3.11
click==7.1.2
rq==1.2.0
rq-dashboard==0.5.1
gunicorn==19.9.0
requests
Pillow
torch==1.10.1
torchvision